Switch build and shared ABI to x86-64

+++ src/common/types.h
@@ -11,8 +11,8 @@ typedef signed short int16_t;
 typedef signed int int32_t;
 typedef signed long long int64_t;
 
-typedef uint32_t size_t;
-typedef uint32_t uintptr_t;
+typedef uint64_t size_t;
+typedef uint64_t uintptr_t;
 typedef uint8_t bool;
 
 #define true 1
